WIP: input transition 작업
