+++ kbblog/profiles/templates/profiles/profile_post_write.html
@@ -5,14 +5,9 @@
     <div>
         <form action="" method="POST">
             {% csrf_token %}
+            
             <div class="posting-write-card common-card">
-                <div class="posting-write-card__title">
-                    {{form.post_title}}
-                </div>
-                <div class="posting-write-card__content">
-                    {{form.media}}
-                    {{form.post_content}}
-                </div>
+                <div id="editor"></div>
             </div>
             <button type="submit" name="post_submit">글쓰기</button> 
         </form>
