+++ kbblog/static/common/devjs/index.ts
@@ -1,0 +1,1 @@
+import './common.ts';
