+++ kbblog/static/profiles/scss/main.css
@@ -144,9 +144,9 @@ button[transition-init=transparent]:hover { background-color: rgba(0, 0, 0, 0.04
 
 .transition-input { height: 40px; transition: border-color 250ms; padding: 0 15px; }
 
-.transition-input:hover { border-color: black; }
+.transition-input--active + .transition-label { transform: translate(15px, -7px) scale(0.8); }
 
-.transition-input:value + .transition-label { transform: translate(15px, -7px) scale(0.8); }
+.transition-input:hover { border-color: black; }
 
 .transition-label { position: absolute; top: 0; left: 0; transition: transform 250ms; transform: translate(15px, 8px) scale(1); background-color: #fff; }
