+++ kbblog/static/profiles/scss/postGlobal.scss
@@ -217,5 +217,10 @@
         width: 100%;
         border: 0;
         border-bottom: 2px solid #333;
+        box-sizing: border-box;
+
+        &--title {
+
+        }
     }
 }
